@@ -14,7 +14,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3",
+        "php": ">=8.0",
         "symfony/process": "^2.8"
     },
     "require-dev": {
